reorganize the common code of the two tests
